Fixed path resolution bugs
