+++ jest.config.js
@@ -1,3 +1,4 @@
+/* eslint-disable */
 module.exports = {
   preset: "ts-jest",
   testEnvironment: "node",
