+++ src/_enums/CheckinItem.ts
@@ -1,0 +1,9 @@
+/**
+ * Fine-grained access control for checkin items
+ */
+export enum CheckinAccessLevel {
+  ALL,
+  SPONSORS_ONLY,
+  PARTICIPANTS_ONLY,
+  ADMINS_ONLY,
+}
