+++ src/_enums/ScheduleItem.ts
@@ -1,0 +1,8 @@
+export enum ScheduleItemPlatform {
+  IN_PERSON,
+  ZOOM,
+  HOPIN,
+  DISCORD,
+  YOUTUBE,
+  OTHER,
+}
