+++ src/_enums/Status.ts
@@ -1,0 +1,9 @@
+/**
+ * Fields of IStatus that do not require recording the authorizer
+ */
+export enum StatusField {
+  VERIFIED = "verified",
+  COMPLETED_PROFILE = "completedProfile",
+  CONFIRMED = "confirmed",
+  DECLINED = "declined",
+}
