+++ src/_enums/TeamRequest.ts
@@ -1,0 +1,11 @@
+export enum TeamRequestType {
+  INVITE,
+  JOIN,
+}
+
+export enum TeamRequestStatus {
+  PENDING,
+  ACCEPTED,
+  DECLINED,
+  CANCELLED,
+}
