+++ src/_types/Email.d.ts
@@ -1,19 +1,5 @@
 import { Document, ObjectId } from "mongoose";
-
-export enum EmailGroup {
-  PARTICIPANTS_VERIFIED,
-  PARTICIPANTS_COMPLETED,
-  PARTICIPANTS_CONFIRMED,
-  SPONSORS,
-  ADMINS,
-}
-
-export enum EmailStatus {
-  QUEUED,
-  SENT,
-  ERROR,
-  DELETED,
-}
+import * as Email from "../_enums/Email";
 
 /**
  * Type for Email model
@@ -22,11 +8,11 @@ export interface IEmail extends Document {
   _id: ObjectId;
   event: ObjectId;
   sender: ObjectId;
-  groups: [EmailGroup];
+  groups: [Email.EmailGroup];
   subject: string;
   body: string;
   sendTime?: Date;
-  status: EmailStatus;
+  status: Email.EmailStatus;
   createdAt: Date;
   updatedAt?: Date;
 }
