+++ src/_types/Status.d.ts
@@ -16,13 +16,3 @@ export interface IStatus extends Document {
   createdAt?: Date;
   updatedAt?: Date;
 }
-
-/**
- * Fields of IStatus that do not require recording the authorizer
- */
-export enum StatusField {
-  VERIFIED = "verified",
-  COMPLETED_PROFILE = "completedProfile",
-  CONFIRMED = "confirmed",
-  DECLINED = "declined",
-}
