+++ src/controllers/UsersController.ts
@@ -1,4 +1,4 @@
-import User from "src/models/User";
+import User from "../models/User";
 import { Request, Response } from "express";
 
 export const getUsers = async (req: Request, res: Response): Promise<void> => {
